Bug 946407: Accept a single unified-memory-report file. r=njn
Using nsIMemoryReporterManager::GetReports to get memory reports from
content processes means that we have a single collection of report
entries instead of one per process, which means a single file.  But this
code has to deal with geckos both before and after that patch.  So,
this change accepts either N individual "memory-report-*" files or one
"unified-memory-report-*" file (and applies this more generally, because
that was easier than making it specific to the main memory report).