Bug 1048819 - Make --sdk-version optionnal, with a default of ICS
With bug 1047350 we introduced a new mandatory option (--sdk-version) in
the script buildinf the FOTA package. However, this is living in B2G/
which is not branched. The calling code that passes the argument lives
in gonk-misc/ which is branched. This makes it possible to have all
repos uptodate and not sync thus causing bug 1048819. We make the option
non mandatory and we default it to ICS.