Bug 1063520 - Compute the list of B2G ELF files instead of hard-coding it g=gsvelto
We explicitely exclude those files because they are part of the update:
checking their SHA1 is useless and will stop the update process because
of the expected mismatch. We instead rely on file to identify the files
in the B2G folder that have a mimetype for shared library or executable.