Bug 1030288: Don't remove directories in |FlashFotaBuilder.zip_filter|
There is an optimization in |FlashFotaBuilder.zip_filter| that
skips unaffected directories when bulding the ZIP file. It only
works correctly if the directory path is included in the list
of FOTA files.

This patch removes the optimization, so that any files can be
included in the ZIP file, even if their base path is not in the
file list.